requirements.txt
absl-py==0.12.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Cython==0.29.22
fastapi==0.63.0
google-auth==1.28.0
google-auth-oauthlib==0.4.4
grpcio==1.36.1
h11==0.12.0
idna==2.10
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.1
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.1.48
Pillow==8.2.0
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.2
six==1.15.0
starlette==0.13.6
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
Werkzeug==1.0.1
